% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump_df_mat_to_file.R
\name{dump_df_mat_to_file}
\alias{dump_df_mat_to_file}
\title{Dump a dataframe/matrix to MS Word file}
\usage{
dump_df_mat_to_file(
  out,
  cols = NULL,
  decimals = 2,
  NA.string = NA,
  fontname = "Arial",
  table.title = NULL,
  table.footnote = NULL,
  print.dir = NULL,
  file.name = NULL
)
}
\arguments{
\item{decimals}{a vector of the number of decimals to round each numeric column to, can also be a scalar
if that number applies to all the numbers; default is 2}

\item{NA.string}{determines how to print out missing (i.e., NA) strings; default is blank}

\item{fontname}{specify the font; default is 'Arial'}

\item{table.title}{pass a character vector specifying the table title}

\item{table.footnote}{pass a character vector specifying a footnote}

\item{print.dir}{pass a file path for the file to be printed out to}

\item{file.name}{pass a name for the file}
}
\value{
outputs table in a MS Word table and prints table to R Viewer
}
\description{
This function takes a dataframe or matrix and dumps it out as a table in a MS Word file. The
table will be printed in the R Viewer as well. The defaults
are such that minimal fussing with the R object is required in order to create
the MS Word table. This was designed to streamline the process of moving from completed analysis
to output.
}
