% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_numbers.R
\name{round_numbers}
\alias{round_numbers}
\title{Round Numbers}
\usage{
round_numbers(out, cols = NULL, decimals = 2, NA.string = NA)
}
\arguments{
\item{decimals}{vector specifies the number of decimals to round each numeric column to; default is 2}
}
\value{
will return a dataframe of character values
}
\description{
Round numbers to a specified number of decimal places
converts numeric values to a character value
used to prepare the numeric values for a table output
}
