% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell_table.R
\name{shell_table}
\alias{shell_table}
\title{Create a Function to make shell tables}
\usage{
shell_table(out = NULL, cols = NULL, decimals = 2, NA.string = NA)
}
\arguments{
\item{cols}{specify which columns of the dataframe you want the numbers
converted to x}

\item{decimals}{how many xx to add}

\item{NA.string}{determines how to print out missing (i.e., NA) strings; default is blank}

\item{numeric.tables}{pass the data frame}
}
\value{
outputs dataframe with character vectors
}
\description{
Create a Function to make shell tables
}
